Tests/LibWeb/Text/expected/input-image.txt

../../Layout/input/120.png loaded
file:///i-do-no-exist-i-swear.png failed
